requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
psutil==5.9.6
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
schedule==1.2.0